refactor: inputView와 OutputView 도입
